OTP Encryption, Decryption fixed and OTP 30s expiration check is also fixed

+++ One Time Password Web Application/OneTimePassWebApp.API/Data/Requests/OTP/OTPVerifyRequest.cs
@@ -5,7 +5,7 @@
         public int UserID { get; set; }
         public DateTime DateTime { get; set; }
         public string EnteredOTP { get; set; }
-        public double ExpireDate { get; set; }
+        public long ExpireDate { get; set; }
         public string OriginalOTP { get; set; }
     }
 }
